Add ElementValidator class; Improve the efficiency of checking riot/element room message history
